Refactor: extract file
